u64-ization

+++ examples/parallel-compute/main.rs
@@ -29,7 +29,7 @@ fn main() {
             // Current thread GPU objects
             let local_cpu_data = (0..size).into_iter().collect::<Vec<u32>>();
             let local_input_buffer = gpgpu::GpuBuffer::from_slice(&FW, &local_cpu_data);
-            let local_output_buffer = gpgpu::GpuBuffer::<u32>::with_capacity(&FW, size as usize);
+            let local_output_buffer = gpgpu::GpuBuffer::<u32>::with_capacity(&FW, size as u64);
 
             let desc = gpgpu::DescriptorSet::default()
                 .bind_buffer(&local_shader_input_buffer, gpgpu::GpuBufferUsage::ReadOnly)
